Flytter valgbareBehandlinger inn i knyttTilSak.js siden det er der det hører til
